# modules/core/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Core Module API
  description: API for the Core Module providing authentication, event management, and status endpoints
  version: 1.0.0
  contact:
    name: Vacancy System
servers:
  - url: /api/core
    description: Base API path
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Error message describing what went wrong
        stack:
          type: string
          description: Stack trace (only in development environment)
          example: Error at line...
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: yourpassword
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          type: object
          properties:
            id:
              type: string
              example: 123456
            name:
              type: string
              example: User Name
            email:
              type: string
              format: email
              example: user@example.com
            role:
              type: string
              enum: [USER, ADMIN]
              example: USER
...
paths:
...
  /auth/login:
    post:
      summary: User Login
      description: Authenticates a user and returns a JWT token
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...